+++ cscglobal-caplugin/CHANGELOG.md
@@ -1,5 +1,6 @@
 v1.1.0  
 - Support for custom fields in enrollment  
+- Support for returning CNAME tokens from enrollment call  
 
 v1.0.1 
 - Fixed issue with SANs not being read correctly.